Move most error handling from Ui to other classes
